Move getInitDataFromPro from content_protection to pssh util.
