+++ lib/dash/content_protection.js
@@ -21,6 +21,7 @@ goog.require('goog.asserts');
 goog.require('shaka.log');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.ManifestParserUtils');
+goog.require('shaka.util.Pssh');
 goog.require('shaka.util.Uint8ArrayUtils');
 goog.require('shaka.util.XmlUtils');
 
@@ -320,63 +321,6 @@ shaka.dash.ContentProtection = class {
     return ContentProtection.getLaurl_(rootElement);
   }
 
-
-  /**
-   * Gets a PlayReady initData from a content protection element
-   * containing a PlayReady Pro Object
-   *
-   * @param {shaka.dash.ContentProtection.Element} element
-   * @return {?Array.<shaka.extern.InitDataOverride>}
-   */
-  static getInitDataFromPro(element) {
-    const proNode = shaka.util.XmlUtils.findChildNS(
-        element.node, 'urn:microsoft:playready', 'pro');
-    if (!proNode) {
-      return null;
-    }
-    const Uint8ArrayUtils = shaka.util.Uint8ArrayUtils;
-    // 'pssh' 4 bytes
-    const psshString = new Uint8Array([
-      0x70, 0x73, 0x73, 0x68,
-    ]);
-    const flags = new Uint8Array([
-      0x00, 0x00, 0x00, 0x00,
-    ]);
-    const systemId = new Uint8Array([
-      0x9a, 0x04, 0xf0, 0x79, 0x98, 0x40, 0x42, 0x86,
-      0xab, 0x92, 0xe6, 0x5b, 0xe0, 0x88, 0x5f, 0x95,
-    ]);
-    const data = Uint8ArrayUtils.fromBase64(proNode.textContent);
-    const dataLength = data.length;
-    const psshSize = 0x4 + 0x4 + 0x4 + systemId.length + 0x4 + dataLength;
-    /** @type {!ArrayBuffer} */
-    const psshBoxBuffer = new ArrayBuffer(psshSize);
-    /** @type {!Uint8Array} */
-    const psshBox = new Uint8Array(psshBoxBuffer);
-    /** @type {!DataView} */
-    const psshData = new DataView(psshBoxBuffer);
-    let byteCursor = 0;
-    psshData.setUint32(byteCursor, psshSize);
-    byteCursor += 0x4;
-    psshBox.set(psshString, byteCursor);
-    byteCursor += 0x4;
-    psshBox.set(flags, byteCursor);
-    byteCursor += 0x4;
-    psshBox.set(systemId, byteCursor);
-    byteCursor += systemId.length;
-    psshData.setUint32(byteCursor, dataLength);
-    byteCursor += 0x4;
-    psshBox.set(data, byteCursor);
-    byteCursor += dataLength;
-    goog.asserts.assert(byteCursor === psshSize,
-        'MS PRO invalid length.');
-    return [{
-      initData: psshBox,
-      initDataType: 'cenc',
-      keyId: element.keyId,
-    }];
-  }
-
   /**
    * Creates DrmInfo objects from the given element.
    *
@@ -402,7 +346,7 @@ shaka.dash.ContentProtection = class {
             !element.init || element.init.length,
             'Init data must be null or non-empty.');
 
-        const proInitData = ContentProtection.getInitDataFromPro(element);
+        const proInitData = shaka.util.Pssh.getInitDataFromPro(element);
         const initData = element.init || defaultInit || proInitData;
         const info = ManifestParserUtils.createDrmInfo(keySystem, initData);
         const licenseParser = licenseUrlParsers.get(keySystem);
