+++ lib/util/pssh.js
@@ -22,6 +22,7 @@ goog.require('shaka.log');
 goog.require('shaka.util.Iterables');
 goog.require('shaka.util.Mp4Parser');
 goog.require('shaka.util.Uint8ArrayUtils');
+goog.require('shaka.util.XmlUtils');
 
 
 /**
@@ -108,5 +109,61 @@ shaka.util.Pssh = class {
     if (box.reader.getPosition() != box.reader.getLength()) {
       shaka.log.warning('Mismatch between box size and data size!');
     }
+  }
+
+  /**
+   * Gets a PlayReady initData from a content protection element
+   * containing a PlayReady Pro Object
+   *
+   * @param {shaka.dash.ContentProtection.Element} element
+   * @return {?Array.<shaka.extern.InitDataOverride>}
+   */
+  static getInitDataFromPro(element) {
+    const proNode = shaka.util.XmlUtils.findChildNS(
+        element.node, 'urn:microsoft:playready', 'pro');
+    if (!proNode) {
+      return null;
+    }
+    const Uint8ArrayUtils = shaka.util.Uint8ArrayUtils;
+    // 'pssh' 4 bytes
+    const psshString = new Uint8Array([
+      0x70, 0x73, 0x73, 0x68,
+    ]);
+    const flags = new Uint8Array([
+      0x00, 0x00, 0x00, 0x00,
+    ]);
+    const systemId = new Uint8Array([
+      0x9a, 0x04, 0xf0, 0x79, 0x98, 0x40, 0x42, 0x86,
+      0xab, 0x92, 0xe6, 0x5b, 0xe0, 0x88, 0x5f, 0x95,
+    ]);
+    const data = Uint8ArrayUtils.fromBase64(proNode.textContent);
+    const dataLength = data.length;
+    const psshSize = 0x4 + 0x4 + 0x4 + systemId.length + 0x4 + dataLength;
+    /** @type {!ArrayBuffer} */
+    const psshBoxBuffer = new ArrayBuffer(psshSize);
+    /** @type {!Uint8Array} */
+    const psshBox = new Uint8Array(psshBoxBuffer);
+    /** @type {!DataView} */
+    const psshData = new DataView(psshBoxBuffer);
+    let byteCursor = 0;
+    psshData.setUint32(byteCursor, psshSize);
+    byteCursor += 0x4;
+    psshBox.set(psshString, byteCursor);
+    byteCursor += 0x4;
+    psshBox.set(flags, byteCursor);
+    byteCursor += 0x4;
+    psshBox.set(systemId, byteCursor);
+    byteCursor += systemId.length;
+    psshData.setUint32(byteCursor, dataLength);
+    byteCursor += 0x4;
+    psshBox.set(data, byteCursor);
+    byteCursor += dataLength;
+    goog.asserts.assert(byteCursor === psshSize,
+        'MS PRO invalid length.');
+    return [{
+      initData: psshBox,
+      initDataType: 'cenc',
+      keyId: element.keyId,
+    }];
   }
 };
